add fast-epub arg, tweak thread pool, error handling for nftw

+++ src/tpool.h
@@ -8,7 +8,7 @@ typedef struct tpool tpool_t;
 
 typedef void (*thread_func_t)(void *arg);
 
-tpool_t *tpool_create(size_t num, void (*cleanup_func)(), int free_arg);
+tpool_t *tpool_create(int num, void (*cleanup_func)(), int free_arg);
 void tpool_start(tpool_t *pool);
 void tpool_destroy(tpool_t *pool);
